==> 02_get_financial_data.ipynb <==
"""
透過 FinMind API 取得台股歷史資料
url = "https://api.finmindtrade.com/api/v4/data"
"""

# 匯入套件
from FinMind.data import FinMindApi
from FinMind.data import DataLoader
import os
import pandas as pd
import numpy as np
import requests

# 讀取 yfinance 股價資訊，用於對照需要抓取的公司代號
df = pd.read_csv("stock_data.csv")

api = FinMindApi()
token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJkYXRlIjoiMjAyNS0xMi0wMiAwMTowOTowMyIsInVzZXJfaWQiOiJKYXNvbkhzdSIsImlwIjoiMS4xNzMuOTUuODgifQ.Fm06qAQzlbQvRofQX9d4IV95gDqMSBeCeH8xCfg9gWg"
url = "https://api.finmindtrade.com/api/v4/data"
headers = {"Authorization": f"Bearer {token}"}
api.login_by_token(token)
# 查看每日限制
print("每日額度：", api.api_usage_limit)

# 已使用
print("已使用：", api.api_usage)

remaining = api.api_usage_limit - api.api_usage
print("剩餘額度：", remaining)

# 建立 blance sheet 財報資料
# 建立資料存放空間
SAVE_PATH = "BalanceSheet.pkl"

# 檢查是否已有財報存放空間
if os.path.exists(SAVE_PATH):
    old_df = pd.read_pickle(SAVE_PATH)
    print("Loaded:", old_df.shape)
else:
    old_df = pd.DataFrame()
    print("No existing file. Starting fresh.")
already_done = set(old_df["stock_id"].unique()) if not old_df.empty else set()

ticker_list = df["ticker"].unique().tolist()
ticker_list = [x.replace('.TW', '') for x in ticker_list]

all_new = []
left_to_do = [t for t in ticker_list if t not in already_done]


print("Total:", len(ticker_list))
print("Already done:", len(already_done))
print("Remaining:", len(left_to_do))

for tid in left_to_do:

    print("Downloading", tid)

    parameter = {
        "dataset": "TaiwanStockBalanceSheet",
        "data_id": tid,
        "start_date": "2015-01-01",
    }

    try:
        resp = requests.get(url, headers=headers, params=parameter).json()

        if "data" in resp and len(resp["data"]) > 0:
            df = pd.DataFrame(resp["data"])
            df["stock_id"] = tid
            all_new.append(df)
            print(tid, "OK")
        else:
            print(tid, "NO DATA")

        # 每抓完一檔，立即存檔（最重要）
        if len(all_new) > 0:
            temp_df = pd.concat([old_df] + all_new, ignore_index=True)
            temp_df.to_pickle(SAVE_PATH)
            print("Saved progress:", temp_df.shape)

    except Exception as e:
        print("Error on", tid, e)

# 抓取 income statement 財報資料
# 建立資料存放空間
SAVE_PATH = "IncomeStatement.pkl"
if os.path.exists(SAVE_PATH):
    old_df = pd.read_pickle(SAVE_PATH)
    print("Loaded:", old_df.shape)
else:
    old_df = pd.DataFrame()
    print("No existing file. Starting fresh.")
already_done = set(old_df["stock_id"].unique()) if not old_df.empty else set()

ticker_list = df["ticker"].unique().tolist()
ticker_list = [x.replace('.TW', '') for x in ticker_list]
all_new = []
left_to_do = [t for t in ticker_list if t not in already_done]

print("Total:", len(ticker_list))
print("Already done:", len(already_done))
print("Remaining:", len(left_to_do))

for tid in left_to_do:

    print("Downloading", tid)

    parameter = {
        "dataset": "TaiwanStockIncomeStatement",
        "data_id": tid,
        "start_date": "2015-01-01",
    }

    try:
        resp = requests.get(url, headers=headers, params=parameter).json()

        if "data" in resp and len(resp["data"]) > 0:
            df = pd.DataFrame(resp["data"])
            df["stock_id"] = tid
            all_new.append(df)
            print(tid, "OK")
        else:
            print(tid, "NO DATA")

        if len(all_new) > 0:
            temp_df = pd.concat([old_df] + all_new, ignore_index=True)
            temp_df.to_pickle(SAVE_PATH)
            print("Saved progress:", temp_df.shape)

    except Exception as e:
        print("Error on", tid, e)

# cash flow statement 財報資料
SAVE_PATH = "CashFlowsStatement.pkl"
df = pd.read_csv("stock_data.csv")

if os.path.exists(SAVE_PATH):
    old_df = pd.read_pickle(SAVE_PATH)
    print("Loaded:", old_df.shape)
else:
    old_df = pd.DataFrame()
    print("No existing file. Starting fresh.")
already_done = set(old_df["stock_id"].unique()) if not old_df.empty else set()

ticker_list = df["ticker"].unique().tolist()
ticker_list = [x.replace('.TW', '') for x in ticker_list]

all_new = []
left_to_do = [t for t in ticker_list if t not in already_done]

print("Total:", len(ticker_list))
print("Already done:", len(already_done))
print("Remaining:", len(left_to_do))

for tid in left_to_do:

    print("Downloading", tid)

    parameter = {
        "dataset": "TaiwanStockCashFlowsStatement",
        "data_id": tid,
        "start_date": "2015-01-01",
    }

    try:
        resp = requests.get(url, headers=headers, params=parameter).json()

        if "data" in resp and len(resp["data"]) > 0:
            df = pd.DataFrame(resp["data"])
            df["stock_id"] = tid
            all_new.append(df)
            print(tid, "OK")
        else:
            print(tid, "NO DATA")

        if len(all_new) > 0:
            temp_df = pd.concat([old_df] + all_new, ignore_index=True)
            temp_df.to_pickle(SAVE_PATH)
            print("Saved progress:", temp_df.shape)

    except Exception as e:
        print("Error on", tid, e)
